renaming SettingList
