--- src/components/settingList/SettingList.spec.js
@@ -1,13 +1,0 @@
-import React from "react"
-import { render, cleanup } from "@testing-library/react"
-import SettingList from "./settingList"
-
-afterEach(cleanup)
-
-describe("SettingList tests", () => {
-	test("displays a SettingList component", () => {
-		// Arrange
-		// Act
-		// Assert
-	})
-})
